Porting to Racket, cleaning up code
Added a Racket directory, started porting some things to Racket. Needs a
bunch of work, but Vincent's framework is there, and exclusive-cond
runs.

Added some TODO's to examples.scrbl; that code should really be loaded
from the scheme/ files, or something.

Made code was mostly in sync with the code in the paper.